refactor: 删除旧的代码
- 删掉了旧的任务管理、账号管理、浏览器上下文管理等
- 将BrowserSessionManager单独移到Manager下，合并原来的类型代码
- 增加 AccountManager 的 cleanup 功能
