+++ electron/main/ipc/app.ts
@@ -1,6 +1,6 @@
 import { app, BrowserWindow, shell } from 'electron'
 import { IPC_CHANNELS } from 'shared/ipcChannels'
-import { accountManager } from '#/managers/AccountManager2'
+import { accountManager } from '#/managers/AccountManager'
 import { typedIpcMainHandle } from '#/utils'
 
 function setupIpcHandlers() {
