--- electron/main/ipc/utils.ts
@@ -1,3 +1,0 @@
-import { accountManager } from '#/managers/AccountManager'
-
-export const currentAccountName = () => accountManager.getActiveAccount().name
