--- electron/main/tasks/autoReply/index.ts
@@ -1,6 +1,0 @@
-export interface AutoReplyConfig {
-  source: 'compass' | 'control'
-  ws?: {
-    port: number
-  }
-}
