--- electron/main/tasks/connection/types.ts
@@ -1,14 +1,0 @@
-import type playwright from 'playwright'
-
-export interface BrowserSession {
-  browser: playwright.Browser
-  context: playwright.BrowserContext
-  page: playwright.Page
-}
-
-export interface BrowserConfig {
-  headless?: boolean
-  storageState?: string
-}
-
-export type StorageState = playwright.BrowserContextOptions['storageState']
